Tentei arrumar o botão de remover, não consegui entende como ele era pra estar aparecendo na lista, não entendi a organização do modulo.
