+++ src/modulos/calcularMedia.js
@@ -1,4 +1,5 @@
-export function calcularMedia(nota1,nota2,nota3,nota4){
-        const soma = (nota1+nota2+nota3+nota4) / 4;
-    return (soma).toFixed(2);
-}
+export function calcularMedia(nota1, nota2, nota3, nota4){
+    let soma = nota1 + nota2 + nota3 + nota4;
+    let media = soma / 4;
+    return media;
+};
